add wait for root-ca certificate

+++ .github/scripts/validate-deploy.sh
@@ -96,6 +96,18 @@ fi
  # exit 1
 #fi
 
+## Check if the icp4a-root-ca is there 
+count=0
+until kubectl get secret icp4a-root-ca -n "${NAMESPACE}" || [[ $count -eq 60 ]]; do
+  echo "Waiting for secret icp4a-root-ca in ${NAMESPACE} COUNTER $count" 
+  count=$((count + 1))
+  sleep 60
+done
+
+if [[ $count -eq 60 ]]; then
+  echo "Timed out waiting for icp4a-root-ca in ${NAMESPACE}"
+  exit 1
+fi
 
 ## Check if the admin.registrykey is there 
 count=0
@@ -105,6 +117,11 @@ until kubectl get secret admin.registrykey -n "${NAMESPACE}" || [[ $count -eq 20
   sleep 15
 done
 
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for admin.registrykey in ${NAMESPACE}"
+  exit 1
+fi
+
 ## Check if the subscription for ibm-automation is there 
 SUBSNAME="ibm-automation"
 count=0
@@ -113,15 +130,13 @@ until kubectl get subs -n "${NAMESPACE}" |grep "${SUBSNAME}" || [[ $count -eq 20
   echo "Waiting for Subscription/${SUBSNAME} in ${NAMESPACE}"
   count=$((count + 1))
   sleep 15
+  exit 1
 done
 
-## Check if the icp4a-root-ca is there 
-count=0
-until kubectl get secret icp4a-root-ca -n "${NAMESPACE}" || [[ $count -eq 40 ]]; do
-  echo "Waiting for secret icp4a-root-ca in ${NAMESPACE} COUNTER $count" 
-  count=$((count + 1))
-  sleep 60
-done
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for Subscription/${SUBSNAME} in ${NAMESPACE}"
+  exit 1
+fi
 
 ## Check if the configmaps is there 
 count=0
@@ -129,14 +144,20 @@ until kubectl get configmaps icp4adeploy-gitops-cp-odm-access-info -n "${NAMESPA
   echo "Waiting for configmaps icp4adeploy-gitops-cp-odm-access-info in ${NAMESPACE} COUNTER $count" 
   count=$((count + 1))
   sleep 40
+  exit 1
 done
+
+if [[ $count -eq 30 ]]; then
+  echo "Timed out waiting for icp4adeploy-gitops-cp-odm-access-info in ${NAMESPACE}"
+  exit 1
+fi
 
 #### Temporary sleep to validate deployment manually
 count=0
 echo "Sleeping for 10 minutes after finding the subscription to manually verify"
 sleep 600
 
-kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
+#kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
 
 cd ..
 rm -rf .testrepo
